refactor: Initial refactor
We break out nested classes into top-level classes. We also move
logging code into a separate package. This is preparation for moving
the CLI code (using commons-cli) into a separate package and then
a separate artifact.
