+++ org.eclipse.transformer.cli/src/main/java/org/eclipse/transformer/jakarta/JakartaTransformer.java
@@ -14,8 +14,8 @@ package org.eclipse.transformer.jakarta;
 import java.util.HashMap;
 import java.util.Map;
 
+import org.eclipse.transformer.AppOption;
 import org.eclipse.transformer.Transformer;
-import org.eclipse.transformer.Transformer.AppOption;
 import org.osgi.annotation.bundle.Header;
 
 @Header(name = "Main-Class", value = "${@class}")
@@ -38,8 +38,8 @@ public class JakartaTransformer {
 	public static final String	DEFAULT_MASTER_TEXT_REFERENCE	= "jakarta-text-master.properties";
 	public static final String	DEFAULT_PER_CLASS_CONSTANT_MASTER_REFERENCE	= "jakarta-per-class-constant-master.properties";
 
-	public static Map<Transformer.AppOption, String> getOptionDefaults() {
-		HashMap<Transformer.AppOption, String> optionDefaults = new HashMap<>();
+	public static Map<AppOption, String> getOptionDefaults() {
+		HashMap<AppOption, String> optionDefaults = new HashMap<>();
 
 		optionDefaults.put(AppOption.RULES_RENAMES, DEFAULT_RENAMES_REFERENCE);
 		optionDefaults.put(AppOption.RULES_VERSIONS, DEFAULT_VERSIONS_REFERENCE);
