+++ org.eclipse.transformer/src/main/java/org/eclipse/transformer/log/package-info.java
@@ -1,0 +1,12 @@
+/********************************************************************************
+ * Copyright (c) Contributors to the Eclipse Foundation
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0, or the Apache License, Version 2.0
+ * which is available at https://www.apache.org/licenses/LICENSE-2.0.
+ *
+ * SPDX-License-Identifier: (EPL-2.0 OR Apache-2.0)
+ ********************************************************************************/
+
+package org.eclipse.transformer.log;
